pizza app improved

+++ class-notes/23-PizzaAPI JWT/src/controllers/auth.js
@@ -9,7 +9,6 @@ const Token = require('../models/token')
 const passwordEncrypt = require('../helpers/passwordEncrypt')
 const jwt = require('jsonwebtoken')
 
-
 module.exports = {
 
     login: async (req, res) => {
@@ -50,46 +49,41 @@ module.exports = {
 
                     /* JWT */
 
-                    accessInfo = {
-                        key : process.env.ACCESS_KEY,
-                        time : "30dm",
-                        data : {
-                            id : user.id,
-                            username : user.username,
-                            email : user.email,
-                            password : user.password,
-                            isAdmin : user.isAdmin,
-                            isActive : user.isActive
-                        }
-
-                    }
-                    refreshInfo = {
-                        key : process.env.REFRESH_KEY,
-                        time : "3d",
-                        data : {
-                            id : user.id,
-                            password : user.password
-                        }
+                    const accessInfo = {
+                        key: process.env.ACCESS_KEY,
+                        time: process.env?.ACCESS_EXP || '30m',
+                        data: {
+                            _id: user._id,
+                            id: user.id,
+                            username: user.username,
+                            email: user.email,
+                            password: user.password,
+                            isActive: user.isActive,
+                            isAdmin: user.isAdmin,
+                        },
                     }
 
-                    // jwt.sign(payload, secretOrPrivateKey, [options, callback])
-                    // jwt.verify(token, secretOrPublicKey, [options, callback]) 
+                    const refreshInfo = {
+                        key: process.env.REFRESH_KEY,
+                        time: process.env?.REFRESH_EXP || '30m',
+                        data: {
+                            id: user.id,
+                            password: user.password // encrypted password
+                        },
+                    }
 
+                    // jwt.sign(access_data, access_key, { expiresIn: '30m' })
                     const accessToken = jwt.sign(accessInfo.data, accessInfo.key, { expiresIn: accessInfo.time })
                     const refreshToken = jwt.sign(refreshInfo.data, refreshInfo.key, { expiresIn: refreshInfo.time })
 
-                    
-
-
                     /* JWT */
-
 
                     res.status(200).send({
                         error: false,
                         token: tokenData.token,
                         bearer: {
-                            access : accessToken,
-                            refresh : refreshToken
+                            access: accessToken,
+                            refresh: refreshToken,
                         },
                         user
                     })
@@ -113,6 +107,51 @@ module.exports = {
 
     },
 
+    refresh: async (req, res) => {
+        /*
+            #swagger.tags = ["Authentication"]
+            #swagger.summary = "JWT: Refresh"
+            #swagger.description = 'Refresh token.'
+        */
+
+        const refreshToken = req.body?.bearer?.refresh
+
+        if (refreshToken) {
+
+            const refreshData = await jwt.verify(refreshToken, process.env.REFRESH_KEY)
+            // console.log(refreshData)
+
+            if (refreshData) {
+
+                const user = await User.findOne({ _id: refreshData.id })
+
+                if (user && user.password == refreshData.password) {
+
+                    res.status(200).send({
+                        error: false,
+                        bearer: {
+                            access: jwt.sign(user.toJSON(), process.env.ACCESS_KEY, { expiresIn: (process.env?.ACCESS_EXP || '30m') })
+                        }
+                    })
+
+                } else {
+
+                    res.errorStatusCode = 401
+                    throw new Error('Wrong id or password.')
+                }
+
+            } else {
+                res.errorStatusCode = 401
+                throw new Error('JWT refresh data is wrong.')
+            }
+
+        } else {
+            res.errorStatusCode = 401
+            throw new Error('Please enter bearer.refresh')
+        }
+
+    },
+
     logout: async (req, res) => {
         /*
             #swagger.tags = ["Authentication"]
@@ -122,13 +161,23 @@ module.exports = {
 
         const auth = req.headers?.authorization // Token ...tokenKey...
         const tokenKey = auth ? auth.split(' ') : null // ['Token', '...tokenKey...']
-        const result = await Token.deleteOne({ token: tokenKey[1] })
 
-        res.send({
-            error: false,
-            message: 'Token deleted. Logout was OK.',
-            result
-        })
+
+        if (tokenKey[0] == 'Token') {
+
+            const result = await Token.deleteOne({ token: tokenKey[1] })
+            res.send({
+                error: false,
+                message: 'Token deleted. Logout was OK.',
+                result
+            })
+
+        } else {
+
+            res.send({
+                error: false,
+                message: 'JWT: No need any process for logout.',
+            })
+        }
     }
-
 }
